LICENSE: change to MIT
This repository contains almost nothing and all of the contents was produced
by Morphbits employees. It's based on public data under permissive licenses
(varios DBs) and just repackages it in some ways. CC licenses are not the best
choice for the code, so better have something more suitable.

We will also import a bit more code here that will do these conversions and
provide a proper package (see #15, #17), this better be properly licensed for
a very generic library containing public data inside (even Apache 2.0 seems
to be inappropriate).

Signed-off-by: Roman Khimov <roman@nspcc.ru>