Makefile: use static revision for openflights data
It's very rarely changed, so it's not a problem. And we can have a bit less
surpises and also close #13 since after this change we no longer have any
simple files to download, they're all coming from github with known revisions.
If Github is to return something different and we can't trust it any additional
checks would be meaningless anyway.

Signed-off-by: Roman Khimov <roman@nspcc.ru>